task2-dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
